refactor: alchemer pydantic
