+++ tests/assets/test_google_directory_assets.py
@@ -34,10 +34,11 @@ def _test_asset(asset: AssetsDefinition):
     assert result.success
     assert (
         result.get_asset_materialization_events()[0]
-        .event_specific_data.materialization.metadata["record_count"]
+        .event_specific_data.materialization.metadata["record_count"]  # type: ignore
         .value
         > 0
     )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
 
 
 def test_asset_google_directory_groups():
