+++ tests/assets/test_google_forms_assets.py
@@ -28,6 +28,7 @@ def _test_asset(asset: AssetsDefinition, partition_key: str | None = None):
         .value
         > 0
     )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
 
 
 def test_asset_google_forms_form():
