+++ tests/assets/test_ldap_assets.py
@@ -21,10 +21,11 @@ def _test_ldap_asset(name, search_base, search_filter):
     assert result.success
     assert (
         result.get_asset_materialization_events()[0]
-        .event_specific_data.materialization.metadata["records"]
+        .event_specific_data.materialization.metadata["records"]  # type: ignore
         .value
         > 0
     )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
 
 
 """ does not work in dev: IP filter
