+++ tests/assets/test_performance_management_assets.py
@@ -20,3 +20,10 @@ def test_outlier_detection():
     )
 
     assert result.success
+    assert (
+        result.get_asset_materialization_events()[0]
+        .event_specific_data.materialization.metadata["record_count"]  # type: ignore
+        .value
+        > 0
+    )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
