+++ tests/assets/test_sftp_assets.py
@@ -76,12 +76,13 @@ def _test_asset(
     )
 
     assert result.success
-    # assert (
-    #     result.get_asset_materialization_events()[0]
-    #     .event_specific_data.materialization.metadata["records"]
-    #     .value
-    #     > 0
-    # )
+    assert (
+        result.get_asset_materialization_events()[0]
+        .event_specific_data.materialization.metadata["records"]  # type: ignore
+        .value
+        > 0
+    )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
 
 
 def test_asset_edplan():
