+++ tests/assets/test_smartrecruiters_assets.py
@@ -25,6 +25,7 @@ def _test_asset(asset_name, report_id):
         .value
         > 0
     )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
 
 
 def test_asset_smartrecruiters_applicants():
