+++ tests/assets/test_tableau_assets.py
@@ -28,3 +28,4 @@ def test_workbook():
         .value
         > 0
     )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
