+++ tests/assets/test_zendesk_assets.py
@@ -23,6 +23,13 @@ def _test_asset(asset: AssetsDefinition, partition_key: str | None = None):
     )
 
     assert result.success
+    assert (
+        result.get_asset_materialization_events()[0]
+        .event_specific_data.materialization.metadata["record_count"]  # type: ignore
+        .value
+        > 0
+    )
+    assert result.get_asset_check_evaluations()[0].metadata.get("extras").text == ""
 
 
 def test_asset_ticket_metrics_archive():
